split popup storage off so it's easier to test
